Bug fixes to make_ext4fs
Fix definition of uuid struct.
Properly set the block_group number in each backup copy of the superblock.
Adjust the info.total_inodes field after rounding up the inodes per block group.
Add the option -t to initialize all the inode tables.  If also specified with
the -s option, use the CHUNK_TYPE_FILL sparse records to initialize them.

+++ ext4_utils/uuid.c
@@ -21,9 +21,10 @@
 #include "sha1.h"
 #include "uuid.h"
 
+/* Definition from RFC-4122 */
 struct uuid {
 	u32 time_low;
-	u32 time_mid;
+	u16 time_mid;
 	u16 time_hi_and_version;
 	u8 clk_seq_hi_res;
 	u8 clk_seq_low;
